+++ tokenize.c
@@ -107,11 +107,11 @@ void expect(char *s)
     token = token->next;
 }
 
-int expect_number()
+long expect_number()
 {
     if (token->kind != TK_NUM)
         error_tok(token, "数ではありません");
-    int val = token->val;
+    long val = token->val;
     token = token->next;
     return val;
 }
